Privitize DownloadService Github client
There's no reason for this to have a public writer interface, and the
page limiting client module is really a generic that can be moved to
it's own file as needed.
