+++ app/services/download_service.rb
@@ -21,11 +21,6 @@ class DownloadService
     github_id: [:id]
   }.freeze
 
-  def page_limiting_client
-    @page_limiting_client ||= user.github_client.dup.extend(PageLimitingOctokitClient)
-  end
-  attr_writer :page_limiting_client
-
   def fetch_notifications(params: {}, max_results: Octobox.config.max_notifications_to_sync)
     client = page_limiting_client
     params[:max_results] = max_results
@@ -45,6 +40,10 @@ class DownloadService
   end
 
   private
+
+  def page_limiting_client
+    @page_limiting_client ||= user.github_client.dup.extend(PageLimitingOctokitClient)
+  end
 
   def fetch_unread_notifications
     headers = {cache_control: %w(no-store no-cache)}
@@ -93,36 +92,5 @@ class DownloadService
     headers = {cache_control: %w(no-store no-cache)}
     notifications = fetch_notifications(params: {all: true, headers: headers})
     process_read_notifications(notifications)
-  end
-end
-
-module PageLimitingOctokitClient
-  def paginate(url, options = {}, &block)
-    under_max_results = -> (data, max_results) {
-      ! max_results || ! data.respond_to?(:size) || data.size < max_results
-    }
-
-    max_results = options.delete(:max_results)
-    opts = parse_query_and_convenience_headers(options.dup)
-
-    if @auto_paginate || @per_page
-      opts[:query][:per_page] ||=  @per_page || (@auto_paginate ? 100 : nil)
-    end
-
-    data = request(:get, url, opts.dup)
-
-    if @auto_paginate
-      while @last_response.rels[:next] && rate_limit.remaining > 0 && under_max_results.call(data, max_results)
-        @last_response = @last_response.rels[:next].get(:headers => opts[:headers])
-        if block_given?
-          yield(data, @last_response)
-        else
-          data.concat(@last_response.data) if @last_response.data.is_a?(Array)
-        end
-      end
-
-    end
-    data = data.first(max_results) if max_results && data.respond_to?(:first)
-    data
   end
 end
