+++ config/application.rb
@@ -22,5 +22,7 @@ Bundler.require(*Rails.groups)
 module Octobox
   class Application < Rails::Application
     require Rails.root.join('lib/octobox')
+
+    config.autoload_paths << Rails.root.join("lib")
   end
 end
